frontend created with vite vue + create reservation functionality

+++ src/main/java/org/example/cinefylast/api/dto/CreateReservationRequest.java
@@ -3,6 +3,10 @@ package org.example.cinefylast.api.dto;
 import java.time.LocalDateTime;
 
 public class CreateReservationRequest {
-    public record CreateRequest(int showtimeId) {}
-    public record Response(Long id, String reservationCode, int showtimeId, LocalDateTime createdAt) {}
+
+    // Request (vom Client an Server)
+    public record CreateRequest(Long showtimeId) {}
+
+    // Response (vom Server an Client)
+    public record Response(Long id, String reservationCode, Long showtimeId, LocalDateTime createdAt) {}
 }
